+++ test/04-spread.test.js
@@ -375,7 +375,6 @@ it ( 'Replace', () => {
                                   },[])
               ;
               expect ( keyList.length ).to.be.equal ( 2 )
-          //  expect () 
  }) // it attach